Create Contract.sol
Changed unit to uint256.

+++ CryptoZombies/Contract.sol
@@ -2,14 +2,14 @@ pragma solidity >=0.5.0 <0.6.0;
 
 contract ZombieFactory {
 
-    event NewZombie(uint zombieId, string name, uint dna);
+    event NewZombie(uint256 zombieId, string name, uint256 dna);
 
-    uint dnaDigits = 16;
-    uint dnaModulus = 10 ** dnaDigits;
+    uint256 dnaDigits = 16;
+    uint256 dnaModulus = 10 ** dnaDigits;
 
     struct Zombie {
         string name;
-        uint dna;
+         uint256  dna;
     }
 
     Zombie[] public zombies;
